Created a CursorTarget class to accept movement input for square target areas.  The class is documented thoroughly.  A few other classes are based on this one.
Created a simple Arrow class that is a drawable WPF vector shape.

Created a CombatPostureDetector to detect a "ready state".  Ideally, a player will enter this state and then the gesture detector will start to detect a swipe.  Right now, it is triggered by bring both hands together.

Created the ICanvasDrawable for creating classes that can be drawn to a given canvas.  In the samples I've seen, this is done by creating a manipulating controls and then adding them to the Canvas through canvas.Children.Add(...) each frame.

Created the ICanvasUpdatable for creating classes that can be given an IEnumerable for points for detecting cursors on the screen.

Created the PathSelectionComponent which draws four arrows on the screen and detects direction selection.

Updated the SimpleSwipeGestureDetector constants a little for better detection.  This in combination with the posture detector should give pretty good results in detecting swings.

I haven't done any of the plumbing yet but the way I intended for everything to work for combat is for the posture detector to wait for the "ready stance", which triggers the gesture detector to start watching and then reporting the result to the game.  Between combat, the PathSelectionComponent will display the possible directions to go (they all don't have to be enabled) and will relay that selection back to the game.  The path selection works really in the testing I've done.  Using the CursorTarget class, we can create pretty much any object on the screen into  a "button".